Organize the code to be packagable
